Lab2/fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
tensorflow-cpu==2.16.1
numpy>=1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
